+++ Web/MoneySaver.Web/Controllers/RecordsController.cs
@@ -6,7 +6,10 @@
     using System.Threading.Tasks;
 
     using Microsoft.AspNetCore.Http;
+    using Microsoft.AspNetCore.Identity;
     using Microsoft.AspNetCore.Mvc;
+    using MoneySaver.Common;
+    using MoneySaver.Data.Models;
     using MoneySaver.Services.Data.Contracts;
     using MoneySaver.Web.ViewModels.Categories;
     using MoneySaver.Web.ViewModels.Records;
@@ -17,15 +20,26 @@
     {
         private readonly IRecordsService recordsService;
         private readonly IWalletsService walletsService;
+        private readonly UserManager<ApplicationUser> userManager;
 
-        public RecordsController(IRecordsService recordsService, IWalletsService walletsService)
+        public RecordsController(IRecordsService recordsService,
+            IWalletsService walletsService, 
+            UserManager<ApplicationUser> userManager)
         {
             this.recordsService = recordsService;
             this.walletsService = walletsService;
+            this.userManager = userManager;
         }
 
         public async Task<IActionResult> Add(int walletId)
         {
+            var user = await this.userManager.GetUserAsync(this.User);
+
+            if (!await this.walletsService.IsUserOwnWalletAsync(user.Id, walletId))
+            {
+                throw new ArgumentException(GlobalConstants.NoPermissionForEditWallet);
+            }
+
             AddRecordViewModel model = new AddRecordViewModel();
 
             var categories = await this.walletsService.GetWalletCategoriesAsync(walletId);
@@ -47,6 +61,13 @@
         [HttpPost]
         public async Task<IActionResult> Add(AddRecordViewModel input)
         {
+            var user = await this.userManager.GetUserAsync(this.User);
+
+            if (!await this.walletsService.IsUserOwnWalletAsync(user.Id, input.WalletId))
+            {
+                throw new ArgumentException(GlobalConstants.NoPermissionForEditWallet);
+            }
+
             var enumValueAsString = input.Type.ToString();
             await this.recordsService.AddAsync(input.CategoryId, input.Description, input.Amount, enumValueAsString, input.CreatedOn);
             return this.Redirect($"/Wallets/Records/{input.WalletId}");
@@ -54,9 +75,11 @@
 
         public async Task<IActionResult> Edit(string id, int walletId)
         {
-            if (id == null)
+            var user = await this.userManager.GetUserAsync(this.User);
+
+            if (!await this.recordsService.IsUserOwnRecordAsync(user.Id, id))
             {
-                return this.Redirect($"/Wallets/Records/{walletId}");
+                throw new ArgumentException(GlobalConstants.NoPermissionForViewOrEditRecord);
             }
 
             var recordDto = await this.recordsService.GetRecordWithAllCategories(id, walletId);
@@ -87,6 +110,13 @@
         [HttpPost]
         public async Task<ActionResult> Edit(EditRecordInputModel input)
         {
+            var user = await this.userManager.GetUserAsync(this.User);
+
+            if (!await this.recordsService.IsUserOwnRecordAsync(user.Id, input.Id))
+            {
+                throw new ArgumentException(GlobalConstants.NoPermissionForViewOrEditRecord);
+            }
+
             await this.recordsService.UpdateRecord(input.Id, input.CategoryId, input.WalletId, input.Description, input.OldAmount, input.Amount, input.Type, input.CreatedOn);
             return this.Redirect($"/Wallets/Records/{input.WalletId}");
         }
@@ -94,7 +124,14 @@
         [HttpPost]
         public async Task<IActionResult> Delete(string id)
         {
-           int walletId = await this.walletsService.GetWalletIdByRecordIdAsync(id);
+            var user = await this.userManager.GetUserAsync(this.User);
+
+            if (!await this.recordsService.IsUserOwnRecordAsync(user.Id, id))
+            {
+                throw new ArgumentException(GlobalConstants.NoPermissionForViewOrEditRecord);
+            }
+
+            int walletId = await this.walletsService.GetWalletIdByRecordIdAsync(id);
            await this.recordsService.RemoveAsync(id);
 
             // return this.RedirectToAction("All", "Records", new { walletId, action = "Submit", submitAll = false });
